Add evtest.c and build system.
This program is being used everywhere already anyway, but most sources refer
to the c file directly. Debian has it as part of the joystick package.
Keep it in git somewhere so at least I don't lose it.

Man page taken from the Ubuntu joystick package.

autotools seems a bit of overkill for evtest, but...

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>