Prevent out-of-bounds access for undefined types/codes
The various arrays that define the names are of the respective type size
(i.e. rel is of size REL_CNT, etc.). This means the size is encoded at
compile time, with no boundary checks at runtime for whether a type or
code exceeds those limits.

Wrap the accesses into those arrays into helper functions that check for the
maximum allowed sizes for types and for the respective event codes.

Reported-by: Stéphane Chatty <chatty@enac.fr>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>